Fix usability issues found in blind user test
- search() auto-normalizes queries (no manual F.normalize needed)
- build_index() accepts model= and tokenizer= to avoid loading twice
- README: document layer indexing convention and search(layers=)
- README: remove manual normalize from example, add model reuse tip
- Fix torch_dtype deprecation warning (use dtype= instead)

Found via usability test: fresh Claude used the library from just the
README. These were the four friction points it reported.

+++ latentlens/models.py
@@ -86,11 +86,11 @@ def load_model(
     # for VLMs and other architectures
     try:
         model = AutoModelForCausalLM.from_pretrained(
-            model_name, torch_dtype=dtype, trust_remote_code=trust_remote_code
+            model_name, dtype=dtype, trust_remote_code=trust_remote_code
         )
     except (ValueError, KeyError):
         model = AutoModel.from_pretrained(
-            model_name, torch_dtype=dtype, trust_remote_code=trust_remote_code
+            model_name, dtype=dtype, trust_remote_code=trust_remote_code
         )
     model = model.to(device).eval()
